feat: 기대평 등록시 버튼 비활성화, 아래로 스크롤 되며 워드클라우드 표시
