+++ src/client/pages/randomEventResultPage/RandomExpectations/RandomExpectations.tsx
@@ -1,8 +1,15 @@
-import React, { ChangeEvent, forwardRef, useState } from "react";
+import React, {
+  ChangeEvent,
+  forwardRef,
+  useEffect,
+  useRef,
+  useState,
+} from "react";
 import { useErrorBoundary } from "react-error-boundary";
 import { postComment } from "../../../api/fetch";
 import Toast from "../../../components/common/Toast/Toast";
 import { ERROR_MSG } from "../../../constants/RandomEventData";
+import WordCloud from "../../../components/mainPage/InfoScreen/WordCloud";
 
 const RandomExpectations = forwardRef<HTMLDivElement>((props, ref) => {
   const [error, setError] = useState<
@@ -11,6 +18,7 @@ const RandomExpectations = forwardRef<HTMLDivElement>((props, ref) => {
   const [expectation, setExpectation] = useState<string>("");
   const [toastKey, setToastKey] = useState(0);
   const { showBoundary } = useErrorBoundary();
+  const wordCloudRef = useRef<HTMLDivElement>(null);
 
   const onClickHandler = async () => {
     if (expectation.length < 20) setError("short");
@@ -33,6 +41,15 @@ const RandomExpectations = forwardRef<HTMLDivElement>((props, ref) => {
     setExpectation(e.target.value);
   };
 
+  useEffect(() => {
+    if (error === "success" && wordCloudRef.current) {
+      wordCloudRef.current.scrollIntoView({
+        behavior: "smooth",
+        block: "start",
+      });
+    }
+  }, [error]);
+
   return (
     <div className="mb-[4.5rem] w-[94rem]" ref={ref}>
       <hr className="h-[1px] bg-gray-400" />
@@ -49,9 +66,9 @@ const RandomExpectations = forwardRef<HTMLDivElement>((props, ref) => {
         ></textarea>
         <button
           onClick={onClickHandler}
-          className="h-[15.375rem] w-[12.25rem] rounded-[1.75rem] bg-primary font-kia-signature-bold text-title-3 text-gray-950 flex-center"
+          className={`h-[15.375rem] w-[12.25rem] rounded-[1.75rem] ${error === "success" ? "pointer-events-none bg-secondary text-gray-300" : "bg-primary"} font-kia-signature-bold text-title-3 text-gray-950 flex-center`}
         >
-          참여하기
+          {error === "success" ? "참여완료" : "참여하기"}
         </button>
         {error && (
           <Toast
@@ -63,6 +80,16 @@ const RandomExpectations = forwardRef<HTMLDivElement>((props, ref) => {
             duration={1000}
             color={error === "success" ? "bg-[#66bb6a]" : undefined}
           />
+        )}
+      </div>
+      <div className="flex-col flex-center" ref={wordCloudRef}>
+        {error === "success" && (
+          <>
+            <p className="my-4 font-kia-signature-bold text-title-4 text-gray-50">
+              작성한 기대평은 워드 클라우드에 반영됩니다
+            </p>
+            <WordCloud></WordCloud>
+          </>
         )}
       </div>
     </div>
